+++ easygui/boxes/base_boxes.py
@@ -699,316 +699,6 @@ def KeyboardListener(event):
                 return
 
 
-# -------------------------------------------------------------------
-# diropenbox
-# -------------------------------------------------------------------
-def diropenbox(msg=None, title=None, default=None):
-    """
-    A dialog to get a directory name.
-    Note that the msg argument, if specified, is ignored.
-
-    Returns the name of a directory, or None if user chose to cancel.
-
-    If the "default" argument specifies a directory name, and that
-    directory exists, then the dialog box will start with that directory.
-
-    :param str msg: the msg to be displayed
-    :param str title: the window title
-    :param str default: starting directory when dialog opens
-    :return: Normalized path selected by user
-    """
-    title = getFileDialogTitle(msg, title)
-    localRoot = Tk()
-    localRoot.withdraw()
-    if not default:
-        default = None
-    f = ut.tk_FileDialog.askdirectory(
-        parent=localRoot, title=title, initialdir=default, initialfile=None
-    )
-    localRoot.quit()
-    if not f:
-        return None
-    return os.path.normpath(f)
-
-
-# -------------------------------------------------------------------
-# getFileDialogTitle
-# -------------------------------------------------------------------
-def getFileDialogTitle(msg, title):
-    """
-    Create nicely-formatted string based on arguments msg and title
-    :param msg: the msg to be displayed
-    :param title: the window title
-    :return: None
-    """
-    if msg and title:
-        return "%s - %s" % (title, msg)
-    if msg and not title:
-        return str(msg)
-    if title and not msg:
-        return str(title)
-    return None  # no message and no title
-
-
-# -------------------------------------------------------------------
-# class FileTypeObject for use with fileopenbox
-# -------------------------------------------------------------------
-class FileTypeObject:
-
-    def __init__(self, filemask):
-        if len(filemask) == 0:
-            raise AssertionError('Filetype argument is empty.')
-
-        self.masks = list()
-
-        if isinstance(filemask, ut.basestring):  # a str or unicode
-            self.initializeFromString(filemask)
-
-        elif isinstance(filemask, list):
-            if len(filemask) < 2:
-                raise AssertionError('Invalid filemask.\n'
-                                     + 'List contains less than 2 members: "{}"'.format(filemask))
-            else:
-                self.name = filemask[-1]
-                self.masks = list(filemask[:-1])
-        else:
-            raise AssertionError('Invalid filemask: "{}"'.format(filemask))
-
-    def __eq__(self, other):
-        if self.name == other.name:
-            return True
-        return False
-
-    def add(self, other):
-        for mask in other.masks:
-            if mask in self.masks:
-                pass
-            else:
-                self.masks.append(mask)
-
-    def toTuple(self):
-        return self.name, tuple(self.masks)
-
-    def isAll(self):
-        if self.name == "All files":
-            return True
-        return False
-
-    def initializeFromString(self, filemask):
-        # remove everything except the extension from the filemask
-        self.ext = os.path.splitext(filemask)[1]
-        if self.ext == "":
-            self.ext = ".*"
-        if self.ext == ".":
-            self.ext = ".*"
-        self.name = self.getName()
-        self.masks = ["*" + self.ext]
-
-    def getName(self):
-        e = self.ext
-        file_types = {".*": "All", ".txt": "Text",
-                      ".py": "Python", ".pyc": "Python", ".xls": "Excel"}
-        if e in file_types:
-            return '{} files'.format(file_types[e])
-        if e.startswith("."):
-            return '{} files'.format(e[1:].upper())
-        return '{} files'.format(e.upper())
-
-
-# -------------------------------------------------------------------
-# fileopenbox
-# -------------------------------------------------------------------
-def fileopenbox(msg=None, title=None, default='*', filetypes=None, multiple=False):
-    """
-    A dialog to get a file name.
-
-    **About the "default" argument**
-
-    The "default" argument specifies a filepath that (normally)
-    contains one or more wildcards.
-    fileopenbox will display only files that match the default filepath.
-    If omitted, defaults to "\*" (all files in the current directory).
-
-    WINDOWS EXAMPLE::
-
-        ...default="c:/myjunk/*.py"
-
-    will open in directory c:\\myjunk\\ and show all Python files.
-
-    WINDOWS EXAMPLE::
-
-        ...default="c:/myjunk/test*.py"
-
-    will open in directory c:\\myjunk\\ and show all Python files
-    whose names begin with "test".
-
-
-    Note that on Windows, fileopenbox automatically changes the path
-    separator to the Windows path separator (backslash).
-
-    **About the "filetypes" argument**
-
-    If specified, it should contain a list of items,
-    where each item is either:
-
-    - a string containing a filemask          # e.g. "\*.txt"
-    - a list of strings, where all of the strings except the last one
-      are filemasks (each beginning with "\*.",
-      such as "\*.txt" for text files, "\*.py" for Python files, etc.).
-      and the last string contains a filetype description
-
-    EXAMPLE::
-
-        filetypes = ["*.css", ["*.htm", "*.html", "HTML files"]  ]
-
-    .. note:: If the filetypes list does not contain ("All files","*"), it will be added.
-
-    If the filetypes list does not contain a filemask that includes
-    the extension of the "default" argument, it will be added.
-    For example, if default="\*abc.py"
-    and no filetypes argument was specified, then
-    "\*.py" will automatically be added to the filetypes argument.
-
-    :param str msg: the msg to be displayed.
-    :param str title: the window title
-    :param str default: filepath with wildcards
-    :param object filetypes: filemasks that a user can choose, e.g. "\*.txt"
-    :param bool multiple: If true, more than one file can be selected
-    :return: the name of a file, or None if user chose to cancel
-    """
-    localRoot = Tk()
-    localRoot.withdraw()
-
-    initialbase, initialfile, initialdir, filetypes = fileboxSetup(
-        default, filetypes)
-
-    # ------------------------------------------------------------
-    # if initialfile contains no wildcards; we don't want an
-    # initial file. It won't be used anyway.
-    # Also: if initialbase is simply "*", we don't want an
-    # initialfile; it is not doing any useful work.
-    # ------------------------------------------------------------
-    if (initialfile.find("*") < 0) and (initialfile.find("?") < 0):
-        initialfile = None
-    elif initialbase == "*":
-        initialfile = None
-
-    func = ut.tk_FileDialog.askopenfilenames if multiple else ut.tk_FileDialog.askopenfilename
-    ret_val = func(parent=localRoot, title=getFileDialogTitle(msg, title), initialdir=initialdir, initialfile=initialfile, filetypes=filetypes
-                   )
-
-    if multiple:
-        f = [os.path.normpath(x) for x in localRoot.tk.splitlist(ret_val)]
-    else:
-        f = os.path.normpath(ret_val)
-
-    localRoot.quit()
-
-    if not f:
-        return None
-    return f
-
-
-# -------------------------------------------------------------------
-# filesavebox
-# -------------------------------------------------------------------
-def filesavebox(msg=None, title=None, default="", filetypes=None):
-    """
-    A file to get the name of a file to save.
-    Returns the name of a file, or None if user chose to cancel.
-
-    The "default" argument should contain a filename (i.e. the
-    current name of the file to be saved).  It may also be empty,
-    or contain a filemask that includes wildcards.
-
-    The "filetypes" argument works like the "filetypes" argument to
-    fileopenbox.
-
-    :param str msg: the msg to be displayed.
-    :param str title: the window title
-    :param str default: default filename to return
-    :param object filetypes: filemasks that a user can choose, e.g. " \*.txt"
-    :return: the name of a file, or None if user chose to cancel
-    """
-
-    localRoot = Tk()
-    localRoot.withdraw()
-
-    initialbase, initialfile, initialdir, filetypes = fileboxSetup(
-        default, filetypes)
-
-    f = ut.tk_FileDialog.asksaveasfilename(parent=localRoot, title=getFileDialogTitle(msg, title), initialfile=initialfile, initialdir=initialdir, filetypes=filetypes
-                                           )
-    localRoot.quit()
-    if not f:
-        return None
-    return os.path.normpath(f)
-
-
-# -------------------------------------------------------------------
-#
-# fileboxSetup
-#
-# -------------------------------------------------------------------
-def fileboxSetup(default, filetypes):
-    if not default:
-        default = os.path.join(".", "*")
-    initialdir, initialfile = os.path.split(default)
-    if not initialdir:
-        initialdir = "."
-    if not initialfile:
-        initialfile = "*"
-    initialbase, initialext = os.path.splitext(initialfile)
-    initialFileTypeObject = FileTypeObject(initialfile)
-
-    allFileTypeObject = FileTypeObject("*")
-    ALL_filetypes_was_specified = False
-
-    if not filetypes:
-        filetypes = list()
-    filetypeObjects = list()
-
-    for filemask in filetypes:
-        fto = FileTypeObject(filemask)
-
-        if fto.isAll():
-            ALL_filetypes_was_specified = True  # remember this
-
-        if fto == initialFileTypeObject:
-            initialFileTypeObject.add(fto)  # add fto to initialFileTypeObject
-        else:
-            filetypeObjects.append(fto)
-
-    # ------------------------------------------------------------------
-    # make sure that the list of filetypes includes the ALL FILES type.
-    # ------------------------------------------------------------------
-    if ALL_filetypes_was_specified:
-        pass
-    elif allFileTypeObject == initialFileTypeObject:
-        pass
-    else:
-        filetypeObjects.insert(0, allFileTypeObject)
-    # ------------------------------------------------------------------
-    # Make sure that the list includes the initialFileTypeObject
-    # in the position in the list that will make it the default.
-    # This changed between Python version 2.5 and 2.6
-    # ------------------------------------------------------------------
-    if len(filetypeObjects) == 0:
-        filetypeObjects.append(initialFileTypeObject)
-
-    if initialFileTypeObject in (filetypeObjects[0], filetypeObjects[-1]):
-        pass
-    else:
-        if ut.runningPython27:
-            filetypeObjects.append(initialFileTypeObject)
-        else:
-            filetypeObjects.insert(0, initialFileTypeObject)
-
-    filetypes = [fto.toTuple() for fto in filetypeObjects]
-
-    return initialbase, initialfile, initialdir, filetypes
-
-
 def __buttonEvent(event=None, buttons=None, virtual_event=None):
     """
     Handle an event that is generated by a person interacting with a button.  It may be a button press
